Changed converter to return a value in range of int32 timestamp
I hardcoded the response for the date of December 13, 1901 because the micro-
service does not take into account hours, minutes, seconds. Before, the
unix timestamp for that date defaulted to the UTC 00:00 timestamp which is
a greater negative value than the value before an int32 overflow. Hardcoding,
while not ideal, does solve the problem of this microservice not returning a
value in the range of int32.